Move get_maxcolwidth() to wily.helper.
